Docker detection

+++ rhel-docker.sh
@@ -6,14 +6,18 @@ if [ "$EUID" -ne 0 ]
 fi
 
 # Docker setup
-echo "Adding Docker Repo"
-dnf config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo
-echo ""
-echo "Updating DNF"
-dnf update -y
-echo ""
-echo "Installing Docker"
-dnf install docker-ce docker-ce-cli containerd.io docker-compose-plugin -y --allowerasing
-echo ""
-echo "Enabling Docker"
-systemctl enable --now docker
+if [ -f "/usr/bin/docker" ]; then
+    echo "Docker is already installed"
+else
+    echo "Adding Docker Repo"
+    dnf config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo
+    echo ""
+    echo "Updating DNF"
+    dnf update -y
+    echo ""
+    echo "Installing Docker"
+    dnf install docker-ce docker-ce-cli containerd.io docker-compose-plugin -y --allowerasing
+    echo ""
+    echo "Enabling Docker"
+    systemctl enable --now docker
+fi
